Moved sound/loot utils.
