+++ src/engine/core/utils/loot.test.ts
@@ -1,12 +1,46 @@
 import { describe, expect, it } from "@jest/globals";
 
-import { transferLoot } from "@/engine/core/utils/object/object_loot";
+import { isObjectWithValuableLoot, transferLoot } from "@/engine/core/utils/loot";
 import { ammo } from "@/engine/lib/constants/items/ammo";
 import { weapons } from "@/engine/lib/constants/items/weapons";
 import { ClientObject, LuaArray } from "@/engine/lib/types";
 import { mockClientGameObject } from "@/fixtures/xray";
 
-describe("object_loot utils", () => {
+describe("loot utils", () => {
+  it("isObjectWithValuableLoot should correctly check object valuable loot", () => {
+    expect(isObjectWithValuableLoot(mockClientGameObject())).toBe(false);
+    expect(
+      isObjectWithValuableLoot(
+        mockClientGameObject({
+          inventory: [
+            ["grenade_f2", mockClientGameObject({ sectionOverride: "grenade_f2" })],
+            ["grenade_f3", mockClientGameObject({ sectionOverride: "grenade_f3" })],
+          ],
+        })
+      )
+    ).toBe(false);
+    expect(
+      isObjectWithValuableLoot(
+        mockClientGameObject({
+          inventory: [
+            ["grenade_f2", mockClientGameObject({ sectionOverride: "grenade_f2" })],
+            [weapons.wpn_ak74u, mockClientGameObject({ sectionOverride: weapons.wpn_ak74u })],
+          ],
+        })
+      )
+    ).toBe(true);
+    expect(
+      isObjectWithValuableLoot(
+        mockClientGameObject({
+          inventory: [
+            [weapons.wpn_ak74u, mockClientGameObject({ sectionOverride: weapons.wpn_ak74u })],
+            [ammo["ammo_5.45x39_fmj"], mockClientGameObject({ sectionOverride: ammo["ammo_5.45x39_fmj"] })],
+          ],
+        })
+      )
+    ).toBe(true);
+  });
+
   it("transferLoot should correctly transfer lootable items", () => {
     const ak74: ClientObject = mockClientGameObject({ sectionOverride: weapons.wpn_ak74u });
     const akAmmo: ClientObject = mockClientGameObject({ sectionOverride: ammo["ammo_5.45x39_ap"] });
