+++ src/engine/core/utils/object/index.ts
@@ -1,7 +1,6 @@
 export * from "@/engine/core/utils/object/object_action";
 export * from "@/engine/core/utils/object/object_check";
 export * from "@/engine/core/utils/object/object_find";
-export * from "@/engine/core/utils/object/object_loot";
 export * from "@/engine/core/utils/object/object_set";
 export * from "@/engine/core/utils/object/object_setup";
 export * from "@/engine/core/utils/object/object_state";
