+++ src/engine/core/utils/sound.ts
@@ -1,5 +1,6 @@
 import { bit_and, snd_type } from "xray16";
 
+import { registry } from "@/engine/core/database";
 import { LuaLogger } from "@/engine/core/utils/logging";
 import { ESoundType } from "@/engine/lib/constants/sound";
 import { ClientObject, TSoundType } from "@/engine/lib/types";
@@ -69,6 +70,16 @@ export function mapSoundMaskToSoundType(soundMask: TSoundType): ESoundType {
  */
 export function isSoundType(heard: TSoundType, expected: TSoundType): boolean {
   return bit_and(heard, expected) === expected;
+}
+
+/**
+ * Check whether is playing sound.
+ *
+ * @param object - client object to check playing
+ * @returns whether currently sound is playing.
+ */
+export function isPlayingSound(object: ClientObject): boolean {
+  return registry.sounds.generic.has(object.id());
 }
 
 /**
